refgactoring as part 1 towards Flask-app-factory
